refactor: redesign migration as transition actions, rollback on failure
Move all migration orchestration (job creation, PVC rebind, cleanup)
from the ensure() sub-step state machine into transition actions:

- BeginFilestoreMigration: scales down, creates temp PVC + rsync Job
  atomically with the phase transition (eliminates double-job race)
- CompleteFilestoreMigration: deletes old PVC, rebinds PV, creates
  final PVC, cleans up
- RollbackFilestoreMigration: on rsync failure or lost job, reverts
  spec.filestore.storageClass, deletes temp PVC + job, transitions
  to Starting

The ensure() method now only keeps deployments at 0.

Removes migrationStep and migrationPvName status fields — no longer
needed since there are no internal sub-steps.
