+++ src/bin/statemachine_diagram.rs
@@ -38,6 +38,7 @@ fn action_name(a: &TransitionAction) -> &'static str {
         TransitionAction::FailBackupJob => "FailBackupJob",
         TransitionAction::BeginFilestoreMigration => "BeginFilestoreMigration",
         TransitionAction::CompleteFilestoreMigration => "CompleteFilestoreMigration",
+        TransitionAction::RollbackFilestoreMigration => "RollbackFilestoreMigration",
     }
 }
