+++ tests/integration/migrate_filestore.rs
@@ -18,26 +18,26 @@ use crate::common::*;
 /// Wait for a batch/v1 Job to appear with a name matching the given prefix.
 async fn wait_for_migration_job(client: &kube::Client, ns: &str, prefix: &str) -> String {
     let jobs: Api<Job> = Api::namespaced(client.clone(), ns);
-    let name = wait_for(TIMEOUT, POLL, || {
-        let jobs = jobs.clone();
-        let pfx = prefix.to_string();
-        async move {
-            if let Ok(list) = jobs.list(&ListParams::default()).await {
-                for job in list.items {
-                    if let Some(ref n) = job.metadata.name {
-                        if n.starts_with(&pfx) {
-                            return true;
+    assert!(
+        wait_for(TIMEOUT, POLL, || {
+            let jobs = jobs.clone();
+            let pfx = prefix.to_string();
+            async move {
+                if let Ok(list) = jobs.list(&ListParams::default()).await {
+                    for job in list.items {
+                        if let Some(ref n) = job.metadata.name {
+                            if n.starts_with(&pfx) {
+                                return true;
+                            }
                         }
                     }
                 }
+                false
             }
-            false
-        }
-    })
-    .await;
-    assert!(name, "migration job with prefix {prefix} never appeared");
-
-    // Return the actual job name.
+        })
+        .await,
+        "migration job with prefix {prefix} never appeared"
+    );
     let list = jobs.list(&ListParams::default()).await.unwrap();
     list.items
         .iter()
@@ -70,7 +70,6 @@ async fn fake_pvc_bound(client: &kube::Client, ns: &str, pvc_name: &str, pv_name
     let pvs: Api<PersistentVolume> = Api::all(client.clone());
     let pvcs: Api<PersistentVolumeClaim> = Api::namespaced(client.clone(), ns);
 
-    // Create the PV if it doesn't exist.
     if pvs.get(pv_name).await.is_err() {
         let pv = PersistentVolume {
             metadata: ObjectMeta {
@@ -97,7 +96,6 @@ async fn fake_pvc_bound(client: &kube::Client, ns: &str, pvc_name: &str, pv_name
         pvs.create(&PostParams::default(), &pv).await.unwrap();
     }
 
-    // Patch PVC spec to bind to PV.
     let spec_patch = json!({"spec": {"volumeName": pv_name}});
     pvcs.patch(
         pvc_name,
@@ -107,7 +105,6 @@ async fn fake_pvc_bound(client: &kube::Client, ns: &str, pvc_name: &str, pv_name
     .await
     .unwrap();
 
-    // Patch PVC status to Bound.
     let status_patch = json!({"status": {"phase": "Bound"}});
     pvcs.patch_status(
         pvc_name,
@@ -126,11 +123,10 @@ async fn migrate_filestore_from_running() {
     let ctx = TestContext::new(name).await;
     let (c, ns) = (&ctx.client, ctx.ns.as_str());
 
-    // Fast-track to Running.
     let ready_handle = fast_track_to_running(&ctx, &format!("{name}-init")).await;
     ready_handle.abort();
 
-    // Fake the original PVC as bound with storageClass "standard".
+    // Fake the original PVC as bound.
     let orig_pvc = format!("{name}-filestore-pvc");
     fake_pvc_bound(c, ns, &orig_pvc, &format!("{name}-pv-old")).await;
 
@@ -143,31 +139,28 @@ async fn migrate_filestore_from_running() {
     )
     .await;
 
-    // Should transition to MigratingFilestore.
+    // BeginFilestoreMigration creates temp PVC + rsync Job atomically with
+    // the phase transition.  By the time we see MigratingFilestore, both exist.
     assert!(
         wait_for_phase(c, ns, name, OdooInstancePhase::MigratingFilestore).await,
         "expected MigratingFilestore phase"
     );
 
-    // Fake both deployments as scaled down (envtest has no kubelet).
+    // Fake deployments scaled down.
     fake_deployment_ready(c, ns, name, 0).await;
     fake_deployment_ready(c, ns, &format!("{name}-cron"), 0).await;
 
-    // Wait for temp PVC to appear.
+    // Fake temp PVC as bound (the transition already created it).
     let temp_pvc = format!("{name}-filestore-pvc-temp");
     wait_for_pvc(c, ns, &temp_pvc).await;
-
-    // Fake temp PVC as bound.
     fake_pvc_bound(c, ns, &temp_pvc, &format!("{name}-pv-new")).await;
 
-    // Wait for rsync job.
+    // Fake rsync job success (the transition already created it).
     let job_name = wait_for_migration_job(c, ns, &format!("{name}-migrate-")).await;
-
-    // Fake job success.
     fake_job_succeeded(c, ns, &job_name).await;
 
-    // Help envtest through the DeletingOldPvc → RebindingPv → Complete flow.
-    // Continuously fake PVC state in the background until migration finishes.
+    // CompleteFilestoreMigration fires: deletes old PVC, rebinds PV, creates
+    // final PVC.  Help envtest through PVC lifecycle in the background.
     let pv_name = format!("{name}-pv-new");
     let assist_handle = {
         let client = c.clone();
@@ -178,7 +171,6 @@ async fn migrate_filestore_from_running() {
         tokio::spawn(async move {
             let pvcs: Api<PersistentVolumeClaim> = Api::namespaced(client.clone(), &ns);
             loop {
-                // Force-delete old PVC if stuck.
                 let _ = pvcs
                     .patch(
                         &orig_pvc,
@@ -186,7 +178,6 @@ async fn migrate_filestore_from_running() {
                         &Patch::Merge(&json!({"metadata": {"finalizers": null}})),
                     )
                     .await;
-                // Force-delete temp PVC if stuck.
                 let _ = pvcs
                     .patch(
                         &temp_pvc,
@@ -194,7 +185,6 @@ async fn migrate_filestore_from_running() {
                         &Patch::Merge(&json!({"metadata": {"finalizers": null}})),
                     )
                     .await;
-                // If final PVC exists, fake it as bound.
                 if pvcs.get(&orig_pvc).await.is_ok() {
                     let _ = pvcs
                         .patch(
@@ -216,7 +206,6 @@ async fn migrate_filestore_from_running() {
         })
     };
 
-    // Should transition to Starting.
     assert!(
         wait_for_phase(c, ns, name, OdooInstancePhase::Starting).await,
         "expected Starting phase after migration"
@@ -240,17 +229,14 @@ async fn migrate_filestore_from_stopped() {
     let ctx = TestContext::new(name).await;
     let (c, ns) = (&ctx.client, ctx.ns.as_str());
 
-    // Fast-track to Running, then stop.
     let ready_handle = fast_track_to_running(&ctx, &format!("{name}-init")).await;
     ready_handle.abort();
 
-    // Fake PVC bound.
     let orig_pvc = format!("{name}-filestore-pvc");
     fake_pvc_bound(c, ns, &orig_pvc, &format!("{name}-pv-old")).await;
 
     // Scale to 0 → Stopped.
     patch_instance_spec(c, ns, name, json!({"replicas": 0})).await;
-    // Fake deployments scaled down.
     fake_deployment_ready(c, ns, name, 0).await;
     fake_deployment_ready(c, ns, &format!("{name}-cron"), 0).await;
     assert!(
@@ -258,7 +244,7 @@ async fn migrate_filestore_from_stopped() {
         "expected Stopped"
     );
 
-    // Change storageClass.
+    // Trigger migration.
     patch_instance_spec(
         c,
         ns,
@@ -272,7 +258,6 @@ async fn migrate_filestore_from_stopped() {
         "expected MigratingFilestore"
     );
 
-    // Fast-forward through migration: temp PVC → rsync → rebind.
     let temp_pvc = format!("{name}-filestore-pvc-temp");
     wait_for_pvc(c, ns, &temp_pvc).await;
     fake_pvc_bound(c, ns, &temp_pvc, &format!("{name}-pv-new2")).await;
@@ -280,7 +265,6 @@ async fn migrate_filestore_from_stopped() {
     let job_name = wait_for_migration_job(c, ns, &format!("{name}-migrate-")).await;
     fake_job_succeeded(c, ns, &job_name).await;
 
-    // Help envtest through the PVC rebind flow with a background assistant.
     let pv_name = format!("{name}-pv-new2");
     let assist_handle = {
         let client = c.clone();
@@ -326,7 +310,6 @@ async fn migrate_filestore_from_stopped() {
         })
     };
 
-    // Should end up Stopped since replicas=0.
     assert!(
         wait_for_phase(c, ns, name, OdooInstancePhase::Stopped).await,
         "expected Stopped after migration with replicas=0"
@@ -334,10 +317,10 @@ async fn migrate_filestore_from_stopped() {
     assist_handle.abort();
 }
 
-// ─── Test 3: Rsync failure retries ──────────────────────────────────────────
+// ─── Test 3: Rsync failure → rollback ───────────────────────────────────────
 
 #[tokio::test]
-async fn migrate_filestore_rsync_failure() {
+async fn migrate_filestore_rsync_failure_rollback() {
     let name = "test-mig-fail";
     let ctx = TestContext::new(name).await;
     let (c, ns) = (&ctx.client, ctx.ns.as_str());
@@ -361,7 +344,6 @@ async fn migrate_filestore_rsync_failure() {
         "expected MigratingFilestore"
     );
 
-    // Fake deployments scaled down.
     fake_deployment_ready(c, ns, name, 0).await;
     fake_deployment_ready(c, ns, &format!("{name}-cron"), 0).await;
 
@@ -369,38 +351,26 @@ async fn migrate_filestore_rsync_failure() {
     wait_for_pvc(c, ns, &temp_pvc).await;
     fake_pvc_bound(c, ns, &temp_pvc, &format!("{name}-pv-new3")).await;
 
-    // First rsync job fails.
+    // Rsync job fails.
     let job_name = wait_for_migration_job(c, ns, &format!("{name}-migrate-")).await;
     fake_job_failed(c, ns, &job_name).await;
 
-    // Should still be in MigratingFilestore. The operator retries by creating
-    // a new job — wait for a different job name to appear (proving retry happened).
-    let api: Api<OdooInstance> = Api::namespaced(c.clone(), ns);
+    // RollbackFilestoreMigration fires: reverts storageClass, cleans up,
+    // transitions to Starting.
     assert!(
-        wait_for(TIMEOUT, POLL, || {
-            let api = api.clone();
-            let n = name.to_string();
-            let old_job = job_name.clone();
-            async move {
-                let current = api
-                    .get_status(&n)
-                    .await
-                    .ok()
-                    .and_then(|i| i.status)
-                    .and_then(|s| s.migration_job_name);
-                // Either cleared (None) or changed to a new job name.
-                current.as_deref() != Some(&old_job)
-            }
-        })
-        .await,
-        "expected migration job to be retried after failure"
+        wait_for_phase(c, ns, name, OdooInstancePhase::Starting).await,
+        "expected Starting after rollback"
     );
-    let inst = api.get_status(name).await.unwrap();
-    assert_eq!(
-        inst.status.as_ref().and_then(|s| s.phase.as_ref()),
-        Some(&OdooInstancePhase::MigratingFilestore),
-        "should remain in MigratingFilestore after rsync failure"
-    );
+
+    // Verify storageClass was reverted to "standard" (the original).
+    let api: Api<OdooInstance> = Api::namespaced(c.clone(), ns);
+    let inst = api.get(name).await.unwrap();
+    let sc = inst
+        .spec
+        .filestore
+        .as_ref()
+        .and_then(|f| f.storage_class.as_deref());
+    assert_eq!(sc, Some("standard"), "storageClass should be reverted");
 }
 
 // ─── Test 4: Not triggered during init ──────────────────────────────────────
@@ -411,12 +381,10 @@ async fn migrate_filestore_not_triggered_during_init() {
     let ctx = TestContext::new(name).await;
     let (c, ns) = (&ctx.client, ctx.ns.as_str());
 
-    // Wait for operator to create PVC, then fake it as bound.
     let orig_pvc = format!("{name}-filestore-pvc");
     wait_for_pvc(c, ns, &orig_pvc).await;
     fake_pvc_bound(c, ns, &orig_pvc, &format!("{name}-pv-old")).await;
 
-    // Create init job to enter Initializing.
     let init_api: Api<odoo_operator::crd::odoo_init_job::OdooInitJob> =
         Api::namespaced(c.clone(), ns);
     let init_job: odoo_operator::crd::odoo_init_job::OdooInitJob = serde_json::from_value(json!({
@@ -436,7 +404,6 @@ async fn migrate_filestore_not_triggered_during_init() {
         "expected Initializing"
     );
 
-    // Change storageClass while Initializing.
     patch_instance_spec(
         c,
         ns,
@@ -445,10 +412,8 @@ async fn migrate_filestore_not_triggered_during_init() {
     )
     .await;
 
-    // Give the operator a few reconcile cycles.
     tokio::time::sleep(Duration::from_secs(3)).await;
 
-    // Should still be Initializing, NOT MigratingFilestore.
     let api: Api<OdooInstance> = Api::namespaced(c.clone(), ns);
     let inst = api.get_status(name).await.unwrap();
     let phase = inst.status.as_ref().and_then(|s| s.phase.as_ref());
